+++ vantage6/vantage6/cli/globals.py
@@ -33,3 +33,9 @@ RABBIT_TIMEOUT = 300
 
 # Default port to start the UI on
 DEFAULT_UI_PORT = 5001
+
+# Location of repository to create new algorithm templates from
+ALGORITHM_TEMPLATE_REPO = "gh:vantage6/v6-algorithm-template.git"
+
+# image to use for diagnostics in `v6 test` commands
+DIAGNOSTICS_IMAGE = "harbor2.vantage6.ai/algorithms/diagnostic"